Fix build for x86_64
target  C: sensors.gmin <= vendor/intel/hardware/iio-sensors/./control.c
vendor/intel/hardware/iio-sensors/./control.c: In function 'acquisition_routine':
vendor/intel/hardware/iio-sensors/./control.c:337:10: error: cast from pointer
to integer of different size [-Werror=pointer-to-int-cast]
  int s = (int) param;
          ^
vendor/intel/hardware/iio-sensors/./control.c: In function
'start_acquisition_thread':
vendor/intel/hardware/iio-sensors/./control.c:453:5: error: cast to pointer
from integer of different size [-Werror=int-to-pointer-cast]
     (void*) s);
     ^

+++ control.c
@@ -340,7 +340,7 @@ static void* acquisition_routine (void* param)
 	 * Bionic does not provide pthread_cancel / pthread_testcancel...
 	 */
 
-	int s = (int) param;
+	int s = (int) (size_t) param;
 	int num_fields;
 	struct sensors_event_t data = {0};
 	int c;
@@ -456,7 +456,7 @@ static void start_acquisition_thread (int s)
 	ret = pthread_create(	&sensor_info[s].acquisition_thread,
 				NULL,
 				acquisition_routine,
-				(void*) s);
+				(void*) (size_t) s);
 }
